requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
frozenlist==1.4.1
idna==3.6
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
tortoise-orm==0.20.0
typing_extensions==4.9.0
yarl==1.9.4
